Added a seperate file for event loop and output
Decided to seperate the event loop logic and take it out of the main
file. Also decided to have a seperate output file which will deal with
all of the output of the program.
